refactor(injection): remove background.js and content_script.js
- remove background.js and content_script.js

- add builds/chromium/inject.js, it only injects sources, nothing more
- add firefox/chrome/content/inject.js, it only injects sources, nothing more
- add maxthon/inject.js, it does nothing :(
- add opera/includes/inject.js, it does nothing :(

This closes #31.

BREAKING CHANGES: this commit simply breaks everything.
Ajax requests made to background.js are now ignored and not handled.
Opera and Maxthon versions do not work at all since sources aren't
being injected.
Same for Safari version as it is not fixed at all.

Needs further fixes which are not included in this commit not to make it
enormous.