refactor(ajax/performRequest): export as an instance, not as a factory
Require `app` module in the implementation instead of accepting it as
an argument to the exported factory.
Moreover, don't return the module as a factory.

Don't return it as a function, return as an object with a single method.
It's much easier to mock it this way in unit tests of other modules.

See #165.