feat(photo): add `downloadAlbumAsZip()`, add it to the actions dropdown
Add `photo.downloadAlbumAsZip()` method which downloads photos from an album,
adds them to a zip archive and saves using `jszip`, `filesaver.js`, and
`blob-polyfill`.

Add saving to a zip as an option to the "Actions" dropdown on album pages.
Disable it on "Photos with ..." pages (vk.com/tag...) because the underlying
API method `photos.get` doesn't support that type of album.

Also upgrade `superagent` to ~1.2.0, mainly for the new `superagent.getXHR()`
method. Superagent still doesn't support `xhr.responseType`, so we wrap `getXHR`
and set the `responseType` manually on the returned xhr in
`ajax.performRequest()`.
See visionmedia/superagent#654.

Closes #175.