feat(ajax): add `cache` option
The `cache` option is a bool. Default: `off`. Set to `on` to cache the response.

The scope of the cache is a window, meaning that you can clear it by closing
the tab and opening it again.

Related to VkOpt/VkOpt#b05fbebf2eee3587c4ca706612f46c0266a48d77 and #169.